test: split models wire construction doubles
